{EQEmu PR 2677} [Experience] Change Exp Calculations to be 64 bit where needed.
Before, would see potential overflows causing Experience Loss messages

![image](https://user-images.githubusercontent.com/109764533/209573703-a708ed3e-dddf-48cd-8fd2-07f6a9678fc1.png)

After, no longer sees these messages, and can crank xp mods up insanely high without it breaking.

![image](https://user-images.githubusercontent.com/109764533/209573682-a7bca849-6798-4e6c-ad7c-b90d3ab93fbb.png)

+++ utils/deprecated/perlxs/groups.h
@@ -1,7 +1,7 @@
 	void	DisbandGroup();
 	bool	IsGroupMember(Mob* client);
 	void	CastGroupSpell(Mob* caster,uint16 spellid);
-	void	SplitExp(uint32 exp, Mob* other);
+	void	SplitExp(uint64 exp, Mob* other);
 	void	GroupMessage(Mob* sender,const char* message);
 	int32	GetTotalGroupDamage(Mob* other);
 	void	SplitMoney(uint32 copper, uint32 silver, uint32 gold, uint32 platinum);
