+++ zone/lua_client.cpp
@@ -251,27 +251,27 @@ void Lua_Client::SetDeity(int v) {
 	self->SetDeity(v);
 }
 
-void Lua_Client::AddEXP(uint32 add_exp) {
+void Lua_Client::AddEXP(uint64 add_exp) {
 	Lua_Safe_Call_Void();
 	self->AddEXP(add_exp);
 }
 
-void Lua_Client::AddEXP(uint32 add_exp, int conlevel) {
+void Lua_Client::AddEXP(uint64 add_exp, int conlevel) {
 	Lua_Safe_Call_Void();
 	self->AddEXP(add_exp, conlevel);
 }
 
-void Lua_Client::AddEXP(uint32 add_exp, int conlevel, bool resexp) {
+void Lua_Client::AddEXP(uint64 add_exp, int conlevel, bool resexp) {
 	Lua_Safe_Call_Void();
 	self->AddEXP(add_exp, conlevel, resexp);
 }
 
-void Lua_Client::SetEXP(uint32 set_exp, uint32 set_aaxp) {
+void Lua_Client::SetEXP(uint64 set_exp, uint64 set_aaxp) {
 	Lua_Safe_Call_Void();
 	self->SetEXP(set_exp, set_aaxp);
 }
 
-void Lua_Client::SetEXP(uint32 set_exp, uint32 set_aaxp, bool resexp) {
+void Lua_Client::SetEXP(uint64 set_exp, uint64 set_aaxp, bool resexp) {
 	Lua_Safe_Call_Void();
 	self->SetEXP(set_exp, set_aaxp, resexp);
 }
@@ -2952,9 +2952,9 @@ luabind::scope lua_register_client() {
 	.def("AddAAPoints", (void(Lua_Client::*)(int))&Lua_Client::AddAAPoints)
 	.def("AddAlternateCurrencyValue", (void(Lua_Client::*)(uint32,int))&Lua_Client::AddAlternateCurrencyValue)
 	.def("AddCrystals", (void(Lua_Client::*)(uint32,uint32))&Lua_Client::AddCrystals)
-	.def("AddEXP", (void(Lua_Client::*)(uint32))&Lua_Client::AddEXP)
-	.def("AddEXP", (void(Lua_Client::*)(uint32,int))&Lua_Client::AddEXP)
-	.def("AddEXP", (void(Lua_Client::*)(uint32,int,bool))&Lua_Client::AddEXP)
+	.def("AddEXP", (void(Lua_Client::*)(uint64))&Lua_Client::AddEXP)
+	.def("AddEXP", (void(Lua_Client::*)(uint64,int))&Lua_Client::AddEXP)
+	.def("AddEXP", (void(Lua_Client::*)(uint64,int,bool))&Lua_Client::AddEXP)
 	.def("AddExpeditionLockout", (void(Lua_Client::*)(std::string, std::string, uint32))&Lua_Client::AddExpeditionLockout)
 	.def("AddExpeditionLockout", (void(Lua_Client::*)(std::string, std::string, uint32, std::string))&Lua_Client::AddExpeditionLockout)
 	.def("AddExpeditionLockoutDuration", (void(Lua_Client::*)(std::string, std::string, int))&Lua_Client::AddExpeditionLockoutDuration)
@@ -3362,8 +3362,8 @@ luabind::scope lua_register_client() {
 	.def("SetDeity", (void(Lua_Client::*)(int))&Lua_Client::SetDeity)
 	.def("SetDuelTarget", (void(Lua_Client::*)(int))&Lua_Client::SetDuelTarget)
 	.def("SetDueling", (void(Lua_Client::*)(bool))&Lua_Client::SetDueling)
-	.def("SetEXP", (void(Lua_Client::*)(uint32,uint32))&Lua_Client::SetEXP)
-	.def("SetEXP", (void(Lua_Client::*)(uint32,uint32,bool))&Lua_Client::SetEXP)
+	.def("SetEXP", (void(Lua_Client::*)(uint64,uint64))&Lua_Client::SetEXP)
+	.def("SetEXP", (void(Lua_Client::*)(uint64,uint64,bool))&Lua_Client::SetEXP)
 	.def("SetEXPEnabled", (void(Lua_Client::*)(bool))&Lua_Client::SetEXPEnabled)
 	.def("SetEXPModifier", (void(Lua_Client::*)(uint32,double))&Lua_Client::SetEXPModifier)
 	.def("SetEXPModifier", (void(Lua_Client::*)(uint32,double,int16))&Lua_Client::SetEXPModifier)
