+++ zone/perl_client.cpp
@@ -219,27 +219,27 @@ void Perl_Client_SetDeity(Client* self, uint32 deity_id) // @categories Account 
 	self->SetDeity(deity_id);
 }
 
-void Perl_Client_AddEXP(Client* self, uint32 add_exp) // @categories Experience and Level
+void Perl_Client_AddEXP(Client* self, uint64 add_exp) // @categories Experience and Level
 {
 	self->AddEXP(add_exp);
 }
 
-void Perl_Client_AddEXP(Client* self, uint32 add_exp, uint8 conlevel) // @categories Experience and Level
+void Perl_Client_AddEXP(Client* self, uint64 add_exp, uint8 conlevel) // @categories Experience and Level
 {
 	self->AddEXP(add_exp, conlevel);
 }
 
-void Perl_Client_AddEXP(Client* self, uint32 add_exp, uint8 conlevel, bool resexp) // @categories Experience and Level
+void Perl_Client_AddEXP(Client* self, uint64 add_exp, uint8 conlevel, bool resexp) // @categories Experience and Level
 {
 	self->AddEXP(add_exp, conlevel, resexp);
 }
 
-void Perl_Client_SetEXP(Client* self, uint32 set_exp, uint32 set_aaxp) // @categories Experience and Level
+void Perl_Client_SetEXP(Client* self, uint64 set_exp, uint64 set_aaxp) // @categories Experience and Level
 {
 	self->SetEXP(set_exp, set_aaxp);
 }
 
-void Perl_Client_SetEXP(Client* self, uint32 set_exp, uint32 set_aaxp, bool resexp) // @categories Experience and Level
+void Perl_Client_SetEXP(Client* self, uint64 set_exp, uint64 set_aaxp, bool resexp) // @categories Experience and Level
 {
 	self->SetEXP(set_exp, set_aaxp, resexp);
 }
@@ -2859,9 +2859,9 @@ void perl_register_client()
 	package.add("AddAAPoints", &Perl_Client_AddAAPoints);
 	package.add("AddAlternateCurrencyValue", &Perl_Client_AddAlternateCurrencyValue);
 	package.add("AddCrystals", &Perl_Client_AddCrystals);
-	package.add("AddEXP", (void(*)(Client*, uint32))&Perl_Client_AddEXP);
-	package.add("AddEXP", (void(*)(Client*, uint32, uint8))&Perl_Client_AddEXP);
-	package.add("AddEXP", (void(*)(Client*, uint32, uint8, bool))&Perl_Client_AddEXP);
+	package.add("AddEXP", (void(*)(Client*, uint64))&Perl_Client_AddEXP);
+	package.add("AddEXP", (void(*)(Client*, uint64, uint8))&Perl_Client_AddEXP);
+	package.add("AddEXP", (void(*)(Client*, uint64, uint8, bool))&Perl_Client_AddEXP);
 	package.add("AddExpeditionLockout", (void(*)(Client*, std::string, std::string, uint32))&Perl_Client_AddExpeditionLockout);
 	package.add("AddExpeditionLockout", (void(*)(Client*, std::string, std::string, uint32, std::string))&Perl_Client_AddExpeditionLockout);
 	package.add("AddExpeditionLockoutDuration", (void(*)(Client*, std::string, std::string, int))&Perl_Client_AddExpeditionLockoutDuration);
@@ -3274,8 +3274,8 @@ void perl_register_client()
 	package.add("SetDeity", &Perl_Client_SetDeity);
 	package.add("SetDuelTarget", &Perl_Client_SetDuelTarget);
 	package.add("SetDueling", &Perl_Client_SetDueling);
-	package.add("SetEXP", (void(*)(Client*, uint32, uint32))&Perl_Client_SetEXP);
-	package.add("SetEXP", (void(*)(Client*, uint32, uint32, bool))&Perl_Client_SetEXP);
+	package.add("SetEXP", (void(*)(Client*, uint64, uint64))&Perl_Client_SetEXP);
+	package.add("SetEXP", (void(*)(Client*, uint64, uint64, bool))&Perl_Client_SetEXP);
 	package.add("SetEXPModifier", (void(*)(Client*, uint32, double))&Perl_Client_SetEXPModifier);
 	package.add("SetEXPModifier", (void(*)(Client*, uint32, double, int16))&Perl_Client_SetEXPModifier);
 	package.add("SetEbonCrystals", &Perl_Client_SetEbonCrystals);
